feat(core): 增加新的 KookVoiceChannel 来描述一个语音频道的信息
