+++ simbot-component-kook-core/src/commonMain/kotlin/love/forte/simbot/component/kook/KookVoiceChannel.kt
@@ -1,0 +1,31 @@
+/*
+ *     Copyright (c) 2025. ForteScarlet.
+ *
+ *     This file is part of simbot-component-kook.
+ *
+ *     simbot-component-kook is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     simbot-component-kook is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *     GNU Lesser General Public License for more details.
+ *
+ *     You should have received a copy of the GNU Lesser General Public License
+ *     along with simbot-component-kook,
+ *     If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package love.forte.simbot.component.kook
+
+/**
+ * 一个 KOOK 中的语音子频道。
+ *
+ * @since 4.2.0
+ * @author ForteScarlet
+ */
+public interface KookVoiceChannel : KookChatChannel {
+    // TODO 相关整合
+}
